- helper functions sepered into directory
